Split pong.py into modules
