+++ constants.py
@@ -1,0 +1,8 @@
+WIDTH, HEIGHT = 640, 480
+PADDLE_WIDTH, PADDLE_HEIGHT = 10, 60
+BALL_SIZE = 10
+PADDLE_SPEED = 5
+BALL_SPEED_X, BALL_SPEED_Y = 4, 4
+
+WHITE = (255, 255, 255)
+BLACK = (0, 0, 0)
