+++ pong.py
@@ -1,50 +1,9 @@
 import pygame
 import sys
 
-# Game constants
-WIDTH, HEIGHT = 640, 480
-PADDLE_WIDTH, PADDLE_HEIGHT = 10, 60
-BALL_SIZE = 10
-PADDLE_SPEED = 5
-BALL_SPEED_X, BALL_SPEED_Y = 4, 4
-
-WHITE = (255, 255, 255)
-BLACK = (0, 0, 0)
-
-pygame.init()
-screen = pygame.display.set_mode((WIDTH, HEIGHT))
-pygame.display.set_caption("Pong")
-clock = pygame.time.Clock()
-font = pygame.font.Font(None, 36)
-
-# Entities
-player = pygame.Rect(20, HEIGHT // 2 - PADDLE_HEIGHT // 2, PADDLE_WIDTH, PADDLE_HEIGHT)
-enemy = pygame.Rect(WIDTH - 30, HEIGHT // 2 - PADDLE_HEIGHT // 2, PADDLE_WIDTH, PADDLE_HEIGHT)
-ball = pygame.Rect(WIDTH // 2 - BALL_SIZE // 2, HEIGHT // 2 - BALL_SIZE // 2, BALL_SIZE, BALL_SIZE)
-ball_vel = pygame.Vector2(BALL_SPEED_X, BALL_SPEED_Y)
-
-def draw_start_screen():
-    screen.fill(BLACK)
-    title = font.render("PONG", True, WHITE)
-    prompt = font.render("Press SPACE to start", True, WHITE)
-    screen.blit(title, (WIDTH // 2 - title.get_width() // 2, HEIGHT // 3))
-    screen.blit(prompt, (WIDTH // 2 - prompt.get_width() // 2, HEIGHT // 2))
-    pygame.display.flip()
-
-
-def reset_ball():
-    ball.center = (WIDTH // 2, HEIGHT // 2)
-    ball_vel.x = BALL_SPEED_X if ball_vel.x > 0 else -BALL_SPEED_X
-    ball_vel.y = BALL_SPEED_Y
-
-
-def update_enemy():
-    # Simple NPC AI to follow the ball
-    if enemy.centery < ball.centery:
-        enemy.y += PADDLE_SPEED
-    elif enemy.centery > ball.centery:
-        enemy.y -= PADDLE_SPEED
-    enemy.clamp_ip(pygame.Rect(0, 0, WIDTH, HEIGHT))
+from constants import WIDTH, HEIGHT, PADDLE_SPEED, WHITE, BLACK
+from entities import screen, clock, player, enemy, ball, ball_vel
+from utils import draw_start_screen, reset_ball, update_enemy
 
 
 def main():
